+++ model/modules/features/__init__.py
@@ -2,7 +2,6 @@ import argparse
 from collections.abc import Callable
 from dataclasses import dataclass
 
-from .composed import ComposedFeatures
 from .full_threats import FullThreats
 from .halfka_v2_hm import HalfKav2Hm
 from .input_feature import InputFeature
@@ -50,7 +49,6 @@ def add_feature_args(parser: argparse.ArgumentParser) -> None:
 
 
 __all__ = [
-    "ComposedFeatures",
     "HalfKav2Hm",
     "FullThreats",
     "InputFeature",
